pyproject.toml
[project]
name = "postmark-email-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.1",
    "fastapi[standard]>=0.115.12",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.27.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "pytest>=8.4.0",
    "requests>=2.32.3",
    "pydantic-ai[logfire]>=0.2.14",
    "google-adk>=1.1.1",
    "ruff>=0.11.12",
    "asyncpg>=0.30.0",
    "gunicorn>=23.0.0",
    "pytest-asyncio>=1.0.0",
]
